requirements.txt
django>=3.2.12
djangorestframework>=3.9.0
Pillow>=9.0.1
requests>=2.27.1
pycryptodome>=3.14.1
cryptography>=36.0.2
firebase_admin>=5.2.0
fcm_django>=1.0.8
celery>=5.2.3